# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Paquito Coach API",
    "version": "0.0.0-dev",
    "description": "API REST del MVP. Autenticación por cookie de sesión."
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "setupCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "studioName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  }
                },
                "required": [
                  "setupCode",
                  "studioName",
                  "name",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "coach",
                        "client"
                      ]
                    },
                    "studio": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "clientStatus": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "invited",
                        "pending",
                        "active",
                        "archived",
                        "no_account"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "studio",
                    "clientStatus"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 200
                  },
                  "joinCode": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 20
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  },
                  "healthDataConsent": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password",
                  "healthDataConsent"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "coach",
                        "client"
                      ]
                    },
                    "studio": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "clientStatus": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "invited",
                        "pending",
                        "active",
                        "archived",
                        "no_account"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "studio",
                    "clientStatus"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clients": {
      "get": {
        "tags": [
          "clientes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "invited",
                "pending",
                "active",
                "archived",
                "no_account"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "userId": {
                        "nullable": true,
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "nullable": true,
                        "type": "string"
                      },
                      "phone": {
                        "nullable": true,
                        "type": "string"
                      },
                      "birthDate": {
                        "nullable": true,
                        "type": "string"
                      },
                      "goal": {
                        "nullable": true,
                        "type": "string"
                      },
                      "healthNotes": {
                        "nullable": true,
                        "type": "string"
                      },
                      "privateNotes": {
                        "nullable": true,
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "invited",
                          "pending",
                          "active",
                          "archived",
                          "no_account"
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "userId",
                      "name",
                      "email",
                      "phone",
                      "birthDate",
                      "goal",
                      "healthNotes",
                      "privateNotes",
                      "tags",
                      "status",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clientes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "email": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "phone": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 30
                  },
                  "birthDate": {
                    "nullable": true,
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "goal": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 500
                  },
                  "healthNotes": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 4000
                  },
                  "privateNotes": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 4000
                  },
                  "tags": {
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 30
                    }
                  },
                  "invite": {
                    "default": true,
                    "type": "boolean"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "client": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "userId": {
                          "nullable": true,
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "nullable": true,
                          "type": "string"
                        },
                        "phone": {
                          "nullable": true,
                          "type": "string"
                        },
                        "birthDate": {
                          "nullable": true,
                          "type": "string"
                        },
                        "goal": {
                          "nullable": true,
                          "type": "string"
                        },
                        "healthNotes": {
                          "nullable": true,
                          "type": "string"
                        },
                        "privateNotes": {
                          "nullable": true,
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "invited",
                            "pending",
                            "active",
                            "archived",
                            "no_account"
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "name",
                        "email",
                        "phone",
                        "birthDate",
                        "goal",
                        "healthNotes",
                        "privateNotes",
                        "tags",
                        "status",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "invite": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "url",
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "client",
                    "invite"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}